requirements.txt
absl-py==0.12.0
attrs==21.2.0
chex==0.1.0
cffi==1.14.0
Cython==0.29.22
cloudpickle==1.6.0
dm-haiku==0.0.6
dm-tree==0.1.5
ffmpeg-python==0.2.0
glfw==2.1.0
gym==0.18.0
gitdb==4.0.9
GitPython==3.1.24
matplotlib==3.4.1
protobuf==3.19.0
imageio==2.1.2
jax==0.2.26
jsonschema==3.2.0
lockfile==0.12.2
neptune-client==0.14.0
numpy==1.20.3
opencv-python==4.5.1.*
optax==0.1.0
psutil==5.7.2
pytest
rlax==0.0.4
scipy==1.6.2
six==1.15.0
tensorboardX==2.1
tensorflow==2.7.0
tensorflow_probability==0.16.0
trfl==1.2.0
